Extract timestamp to own header
